@@ -96,7 +96,10 @@ impl ObjectStorage {
         key
     }
 
-    pub async fn object_mut<T: Any + Send>(&self, key: Key<T>) -> impl DerefMut<Target = T> + Send {
+    pub async fn object_mut<T: Any + Send>(
+        &self,
+        key: Key<T>,
+    ) -> impl DerefMut<Target = T> + Send + use<T> {
         let obj = self.storage.get(&key.key).unwrap();
         let guard = obj.clone().lock_owned().await;
         tokio::sync::OwnedMutexGuard::map(guard, |e: &mut (dyn Any + Send)| {
@@ -107,7 +110,7 @@ impl ObjectStorage {
     pub fn object_mut_blocking<T: Any + Send>(
         &self,
         key: Key<T>,
-    ) -> impl DerefMut<Target = T> + Send {
+    ) -> impl DerefMut<Target = T> + Send + use<T> {
         let obj = self.storage.get(&key.key).unwrap();
         let guard = obj.clone().blocking_lock_owned();
         tokio::sync::OwnedMutexGuard::map(guard, |e: &mut (dyn Any + Send)| {
@@ -141,14 +144,17 @@ impl SessionState {
         self.object_storage.blocking_lock().store_object(obj)
     }
 
-    pub async fn object_mut<T: Any + Send>(&self, key: Key<T>) -> impl DerefMut<Target = T> + Send {
+    pub async fn object_mut<T: Any + Send>(
+        &self,
+        key: Key<T>,
+    ) -> impl DerefMut<Target = T> + Send + use<T> {
         self.object_storage.lock().await.object_mut(key).await
     }
 
     pub fn object_mut_blocking<T: Any + Send>(
         &self,
         key: Key<T>,
-    ) -> impl DerefMut<Target = T> + Send {
+    ) -> impl DerefMut<Target = T> + Send + use<T> {
         self.object_storage.blocking_lock().object_mut_blocking(key)
     }
 
@@ -158,7 +164,10 @@ impl SessionState {
         key
     }
 
-    pub fn session_blocking(&self, sid: Key<Session>) -> impl DerefMut<Target = Session> + Send {
+    pub fn session_blocking(
+        &self,
+        sid: Key<Session>,
+    ) -> impl DerefMut<Target = Session> + Send + use<> {
         self.object_mut_blocking(sid)
     }
 
